requirements.txt
beautifulsoup4==4.9.0
selenium==3.14.0
absl-py==0.9.0
altair==4.1.0
appdirs==1.4.3
appnope==0.1.0
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
base58==2.0.1
beautifulsoup4==4.9.0
bleach==3.1.4
blinker==1.4
boto3==1.14.5
botocore==1.17.5
branca==0.4.1
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
colorama==0.4.3
configparser==5.0.0
constantly==15.1.0
crayons==0.3.0
cryptography==2.9.2
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
densenet-pytorch==0.2.0
descartes==1.1.0
distlib==0.3.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
fake-useragent==0.1.11
filelock==3.0.12
Fiona==1.8.13.post1
Flask==1.1.2
folium==0.11.0
future==0.18.2
gast==0.3.3
geopandas==0.7.0
google-auth==1.16.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
happycowler==0.2.4
hyperlink==19.0.0
idna==2.9
imageio==2.8.0
imutils==0.5.3
incremental==17.5.0
inflection==0.4.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
more-itertools==8.3.0
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parsel==1.6.0
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pinterest-api==0.0.8
prometheus-client==0.7.1
prompt-toolkit==3.0.5
Protego==0.1.16
protobuf==3.12.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py3-pinterest==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.4.0b2
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj==2.6.1.post1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
Quandl==3.5.0
queuelib==1.5.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Scrapy==2.1.0
seaborn==0.10.1
selectorlib==0.16.0
selenium==3.14.0
Send2Trash==1.5.0
service-identity==18.1.0
Shapely==1.7.0
six==1.14.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.6
streamlit==0.61.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tifffile==2020.6.3
toml==0.10.1
toolz==0.10.0
torch==1.5.0
torchvision==0.6.0
tornado==5.1.1
tqdm==4.46.0
traitlets==4.3.3
Twisted==20.3.0
tzlocal==2.1
urllib3==1.25.9
validators==0.15.0
virtualenv==20.0.23
w3lib==1.22.0
watchdog==0.10.2
wcwidth==0.1.9
webdriver-manager==2.5.3
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zope.interface==5.1.0
